analyzer: surface snapshot read failures (#64)
